refactor(core): share language-neutral SCIP decode helpers
Go, Python, Ruby, Rust, and TypeScript decoders each carried private
copies of the same text-format primitives (integer extraction,
whitespace splitting, suffix checks, trailing-char stripping, backtick
removal). Move them into core/scip_decode_common.{h,cpp}; language
symbol policy stays in the owning decoder file. Document the boundary
in core/README.md and docs/core_cpp.md.

Verified: make core-build && make core-test (scip_decode_test,
graph_layers_test both exit 0); clang-format via pre-commit.
